Handle Login Page.
