+++ FrontEnd/src/App.js
@@ -1,12 +1,12 @@
 import React from "react";
-import Login from "./Components/Login";
-import Register from "./Components/Register";
-import Game from "./Components/Game";
-import Lobby2 from "./Components/Lobby2";
+import Login from "./Components/Login/Login";
+import Register from "./Components/Register/Register";
+import Game from "./Components/Game/Game";
+import Lobby2 from "./Components/Lobby2/Lobby2";
 import { createBrowserRouter, RouterProvider, ScrollView, StyleSheet, View } from 'react-router-dom';
 import './index.css';
 import "./App.css";
-import Lobby from "./Components/Lobby";
+import Lobby from "./Components/Lobby/Lobby";
 import Layout from "./Components/Layout/Layout";
 import ProtectedRoutes from "./Components/ProtectedRoutes/ProtectedRoutes";
 import GameBoard from "./Components/Dice/Dice";
