+++ FrontEnd/src/index.css
@@ -19,6 +19,11 @@
   text-align: left;
 }
 
+.snak {
+  width: 100%;
+  height:50vh;
+}
+
 #animation {
   zoom: 10%
 }
