+++ Server/config.js
@@ -1,7 +1,7 @@
 export const config = {
     development: {
         username: 'root',
-        password: 'root',
+        password: '',
         database: 'snackladdergame',
         host: 'localhost',
         dialect: 'mysql',
@@ -9,7 +9,7 @@ export const config = {
     },
     test: {
         username: 'root',
-        password: 'root',
+        password: '',
         database: 'snackladdergame',
         host: 'localhost',
         dialect: 'mysql',
@@ -17,7 +17,7 @@ export const config = {
     },
     production: {
         username: 'root',
-        password: 'root',
+        password: '',
         database: 'snackladdergame',
         host: 'localhost',
         dialect: 'mysql',
